Initial file upload
Ready (just about) to begin writing 'draw' functions. These might need to be reworked a time or two, so I want to be sure that I have a clean backup! This way I can go back and start over if/when I really trump something up (eg. deport code that was useful and generally be racist, a failure, and/or awful. I guess the racist and awful part is probably unlikely for me, but the failing part might happen).